requirements.txt
aiogram>=3.1.0
SQLAlchemy>=2.0.25
aiosqlite>=0.19.0
python-dotenv>=1.0.0
